# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NIAT SPI Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users:
    get:
      operationId: listUsers
      tags: [admin]
      summary: List all users
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [admin]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/users/{id}:
    patch:
      operationId: updateUser
      tags: [admin]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [admin]
      summary: Delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /admin/campuses:
    get:
      operationId: listCampuses
      tags: [admin]
      summary: List campuses
      responses:
        "200":
          description: Campus list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campus"
    post:
      operationId: createCampus
      tags: [admin]
      summary: Create campus
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampusInput"
      responses:
        "201":
          description: Created campus
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campus"
  /admin/campuses/{id}:
    patch:
      operationId: updateCampus
      tags: [admin]
      summary: Update campus
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampusUpdate"
      responses:
        "200":
          description: Updated campus
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campus"
    delete:
      operationId: deleteCampus
      tags: [admin]
      summary: Delete campus
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email, role, isActive, campuses, subjects, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
        campuses:
          type: array
          items:
            type: string
        subjects:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        createdBy:
          type: ["string", "null"]
        lastLoginAt:
          type: ["string", "null"]
        createdAt:
          type: string
    UserInput:
      type: object
      required: [name, email, role, password]
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
        password:
          type: string
        campuses:
          type: array
          items:
            type: string
        subjects:
          type: array
          items:
            type: string
        isActive:
          type: boolean
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
        password:
          type: string
        campuses:
          type: array
          items:
            type: string
        subjects:
          type: array
          items:
            type: string
        isActive:
          type: boolean
    Campus:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        instituteId:
          type: ["string", "null"]
        createdAt:
          type: string
    CampusInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        instituteId:
          type: string
...
    CampusUpdate:
      type: object
      properties:
        name:
          type: string
        instituteId:
          type: string
...